Add files via upload
This Smart Home project integrates various components to create an automated and
interactive living space. The user initiates the system by pressing a push button,
which triggers a doorbell sound and prompts an LCD display to show "Welcome
please enter the pass." The user then enters the passcode using a keypad. If the
passcode is incorrect, the LCD displays "Wrong," and a piezo plays an error sound,
allowing the user to try again.

Once the correct passcode is entered, the main home door opens, and the outdoor
lights turn on, with the LCD displaying "Correct." In the bedroom, an ultrasonic sensor
detects motion, causing the door to open automatically, and the LCD to display
"Motion detected." The door then closes after a few seconds.

The home also features a water pool with a water level sensor. When the user
presses a push button, a 4-digit seven-segment display shows the water level. Lastly,
a fan turns on or off with a push button press.